pyproject.toml
[project]
name = "octopus-photos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt==3.1.7",
    "email-validator>=2.3.0",
    "exifread>=3.5.1",
    "fastapi>=0.129.0",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "bumpver>=2025.1131",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
